pyproject.toml
[tool.poetry]
name = "alfabet-scheduler"
version = "0.1.0"
description = ""
authors = ["TomerEinhorn <tomereinhorn7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
passlib = "^1.7.4"
uvicorn = "^0.28.0"
python-jose = "^3.3.0"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
argon2-cffi = "^23.1.0"
apscheduler = "^3.10.4"
pytest = "^8.1.1"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
